AppServie+AzStorage/requirements.txt
Flask>=1.0,<=1.1.2
azure-storage